Enhance phylogenetic network functionality and improve validation mechanisms
- Updated `cut_edges` function to include additional undirected edges and nodes in the SemiDirectedPhyNetwork example.
- Modified `from_phylozoo_dot` to reflect the addition of a new node in the example.
- Enhanced isomorphism checks with new examples demonstrating non-isomorphic networks due to different node labels.
- Adjusted `SemiDirectedPhyNetwork` examples to include more nodes and edges, improving test coverage.
- Added spacing in `CircularSetOrdering` documentation for clarity.
- Cleaned up edge removal documentation in `DirectedMultiGraph` to remove unnecessary output.
- Updated edge identification in `identify_parallel_edge` to use a more accurate method.
- Improved `MixedMultiGraph` documentation to clarify edge removal behavior.
- Enhanced exception handling in `Partition` to provide clearer error messages.
- Updated string representations in `QuartetProfileSet` and `TripletProfileSet` for better readability.
- Adjusted MSA class to print None for nonexistent sequences instead of raising an error.
- Added spacing in distance calculation documentation for clarity.
- Updated Split and WeightedSplitSystem string representations for better readability.
- Enhanced validation suppression context management to allow nested contexts with different methods.
- Added tests for nested context suppression in validation to ensure correct behavior.

+++ src/phylozoo/core/distance/base.py
@@ -55,7 +55,7 @@ class DistanceMatrix(IOMixin):
     >>> len(dm)
     3
     >>> dm.get_distance('A', 'B')
-    np.float64(1.0)
+    1.0
     >>> # Default labels (0, 1, 2, ...)
     >>> dm2 = DistanceMatrix(matrix)
     >>> dm2.labels
@@ -161,7 +161,7 @@ class DistanceMatrix(IOMixin):
         >>> import numpy as np
         >>> dm = DistanceMatrix(np.array([[0, 1], [1, 0]]), labels=['A', 'B'])
         >>> arr = dm.np_array
-        >>> arr[0, 1]
+        >>> float(arr[0, 1])
         1.0
         """
         return self._matrix
